Added ANTLR-generated artifacts as well as suitable runtime.
This is workaround to deal with dependencies conflict with Android Plugin. To use Effect System, one have to remove ANTLR-runtime jars from ideaSDK/plguins/android/lib and put there ANTLR-runtime that comes with this commit.

This commit will be removed as soon as better way to fix dependency conflict will be found.