Prototype where we're trying to infer some additional info for lambdas bodies, which were called more than once.
This commit contains some experiemnts on inferring additional info for non-EXACTLY_ONCE inlined lambdas
for the sake of smartcasts on vars. Desired behaviour is close to the behaviour with loops, examples can
be found in basic.kt test in effects/smartcasts/onCapturedVars/unknown directory.

The main idea is to treat all vars, captured in in-place lambdas as stable. Then, to prevent erroneous
smarcasts, we have to change data-flow through lambdas that can't be guaranteed to be called <=1 times
(like parameters of collections extension).
We're trying to do this in manner, similar to how loops are working: we get all variables that were
written in this closure and overly conservatively clear any information about them.
There are several issues which weren't adressed in this commit:
- MutableDataFlowInfoForArguments can only add some infromation to the context, not subtract. See
  .getResultInfo(), which always does .and() of initialInfo and resultInfo. Thus, if resultInfo
  says that it doesn't know anything about some variable (which is exactly the case for variables,
  captured and written in non-EXACTLY_ONCE lambdas) and initialInfo knows something, initialInfo wins
  according to .and() semantics.
  Possible fix: rewrite MutableDataFlowInfoForArguments so that resultInfo always include initialInfo,
                therefore eliminating need for .and() in .getResultInfo()
- MutableDataFlowInfoForArguments is built before call resolution, therefore, we don't know effects of
  parent call when we're building data flow and can't change data flow through lambdas accordingly.
  Possible fix: change data flow for ALL lambdas (not just those annotated with effects), and then ban
                some smartcasts for unknown lambdas through variable stability evaluation